Sibling of be6fea8db03f88ebcf39a1cb95b5a1be3b2350cd [skip ci]
skip-checks: true

+++ go/test/endtoend/cluster/vttablet_process.go
@@ -21,7 +21,6 @@ import (
 	"bufio"
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"io"
 	"net/http"
@@ -29,7 +28,6 @@ import (
 	"os/exec"
 	"path"
 	"reflect"
-	"strconv"
 	"strings"
 	"syscall"
 	"testing"
@@ -74,7 +72,6 @@ type VttabletProcess struct {
 	VreplicationTabletType      string
 	DbFlavor                    string
 	Charset                     string
-	ConsolidationsURL           string
 
 	//Extra Args to be set before starting the vttablet process
 	ExtraArgs []string
@@ -88,43 +85,43 @@ func (vttablet *VttabletProcess) Setup() (err error) {
 
 	vttablet.proc = exec.Command(
 		vttablet.Binary,
-		"--topo_implementation", vttablet.CommonArg.TopoImplementation,
-		"--topo_global_server_address", vttablet.CommonArg.TopoGlobalAddress,
-		"--topo_global_root", vttablet.CommonArg.TopoGlobalRoot,
-		"--log_queries_to_file", vttablet.FileToLogQueries,
-		"--tablet-path", vttablet.TabletPath,
-		"--port", fmt.Sprintf("%d", vttablet.Port),
-		"--grpc_port", fmt.Sprintf("%d", vttablet.GrpcPort),
-		"--init_shard", vttablet.Shard,
-		"--log_dir", vttablet.LogDir,
-		"--tablet_hostname", vttablet.TabletHostname,
-		"--init_keyspace", vttablet.Keyspace,
-		"--init_tablet_type", vttablet.TabletType,
-		"--health_check_interval", fmt.Sprintf("%ds", vttablet.HealthCheckInterval),
-		"--enable_replication_reporter",
-		"--backup_storage_implementation", vttablet.BackupStorageImplementation,
-		"--file_backup_storage_root", vttablet.FileBackupStorageRoot,
-		"--service_map", vttablet.ServiceMap,
-		"--vtctld_addr", vttablet.VtctldAddress,
-		"--vtctld_addr", vttablet.VtctldAddress,
-		"--vreplication_tablet_type", vttablet.VreplicationTabletType,
-		"--db_charset", vttablet.Charset,
+		"-topo_implementation", vttablet.CommonArg.TopoImplementation,
+		"-topo_global_server_address", vttablet.CommonArg.TopoGlobalAddress,
+		"-topo_global_root", vttablet.CommonArg.TopoGlobalRoot,
+		"-log_queries_to_file", vttablet.FileToLogQueries,
+		"-tablet-path", vttablet.TabletPath,
+		"-port", fmt.Sprintf("%d", vttablet.Port),
+		"-grpc_port", fmt.Sprintf("%d", vttablet.GrpcPort),
+		"-init_shard", vttablet.Shard,
+		"-log_dir", vttablet.LogDir,
+		"-tablet_hostname", vttablet.TabletHostname,
+		"-init_keyspace", vttablet.Keyspace,
+		"-init_tablet_type", vttablet.TabletType,
+		"-health_check_interval", fmt.Sprintf("%ds", vttablet.HealthCheckInterval),
+		"-enable_replication_reporter",
+		"-backup_storage_implementation", vttablet.BackupStorageImplementation,
+		"-file_backup_storage_root", vttablet.FileBackupStorageRoot,
+		"-service_map", vttablet.ServiceMap,
+		"-vtctld_addr", vttablet.VtctldAddress,
+		"-vtctld_addr", vttablet.VtctldAddress,
+		"-vreplication_tablet_type", vttablet.VreplicationTabletType,
+		"-db_charset", vttablet.Charset,
 	)
 	if *isCoverage {
-		vttablet.proc.Args = append(vttablet.proc.Args, "--test.coverprofile="+getCoveragePath("vttablet.out"))
+		vttablet.proc.Args = append(vttablet.proc.Args, "-test.coverprofile="+getCoveragePath("vttablet.out"))
 	}
 	if *PerfTest {
-		vttablet.proc.Args = append(vttablet.proc.Args, "--pprof", fmt.Sprintf("cpu,waitSig,path=vttablet_pprof_%s", vttablet.Name))
+		vttablet.proc.Args = append(vttablet.proc.Args, "-pprof", fmt.Sprintf("cpu,waitSig,path=vttablet_pprof_%s", vttablet.Name))
 	}
 
 	if vttablet.SupportsBackup {
-		vttablet.proc.Args = append(vttablet.proc.Args, "--restore_from_backup")
+		vttablet.proc.Args = append(vttablet.proc.Args, "-restore_from_backup")
 	}
 	if vttablet.EnableSemiSync {
-		vttablet.proc.Args = append(vttablet.proc.Args, "--enable_semi_sync")
+		vttablet.proc.Args = append(vttablet.proc.Args, "-enable_semi_sync")
 	}
 	if vttablet.DbFlavor != "" {
-		vttablet.proc.Args = append(vttablet.proc.Args, fmt.Sprintf("--db_flavor=%s", vttablet.DbFlavor))
+		vttablet.proc.Args = append(vttablet.proc.Args, fmt.Sprintf("-db_flavor=%s", vttablet.DbFlavor))
 	}
 
 	vttablet.proc.Args = append(vttablet.proc.Args, vttablet.ExtraArgs...)
@@ -145,7 +142,6 @@ func (vttablet *VttabletProcess) Setup() (err error) {
 	go func() {
 		if vttablet.proc != nil {
 			vttablet.exit <- vttablet.proc.Wait()
-			close(vttablet.exit)
 		}
 	}()
 
@@ -168,24 +164,22 @@ func (vttablet *VttabletProcess) GetStatus() string {
 	if err != nil {
 		return ""
 	}
-	defer resp.Body.Close()
 	if resp.StatusCode == 200 {
 		respByte, _ := io.ReadAll(resp.Body)
+		defer resp.Body.Close()
 		return string(respByte)
 	}
 	return ""
 }
 
 // GetVars gets the debug vars as map
-func (vttablet *VttabletProcess) GetVars() map[string]any {
+func (vttablet *VttabletProcess) GetVars() map[string]interface{} {
 	resp, err := http.Get(vttablet.VerifyURL)
 	if err != nil {
 		return nil
 	}
-	defer resp.Body.Close()
-
 	if resp.StatusCode == 200 {
-		resultMap := make(map[string]any)
+		resultMap := make(map[string]interface{})
 		respByte, _ := io.ReadAll(resp.Body)
 		err := json.Unmarshal(respByte, &resultMap)
 		if err != nil {
@@ -202,45 +196,8 @@ func (vttablet *VttabletProcess) GetStatusDetails() string {
 	if err != nil {
 		return fmt.Sprintf("Status details failed: %v", err.Error())
 	}
-	defer resp.Body.Close()
-
 	respByte, _ := io.ReadAll(resp.Body)
 	return string(respByte)
-}
-
-// GetConsolidations gets consolidations
-func (vttablet *VttabletProcess) GetConsolidations() (map[string]int, error) {
-	resp, err := http.Get(vttablet.ConsolidationsURL)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get consolidations: %v", err)
-	}
-	defer resp.Body.Close()
-
-	result := make(map[string]int)
-
-	scanner := bufio.NewScanner(resp.Body)
-	for scanner.Scan() {
-		line := scanner.Text()
-		splits := strings.SplitN(line, ":", 2)
-		if len(splits) != 2 {
-			return nil, fmt.Errorf("failed to split consolidations line: %v", err)
-		}
-		// Discard "Length: [N]" lines.
-		if splits[0] == "Length" {
-			continue
-		}
-		countS := splits[0]
-		countI64, err := strconv.ParseInt(countS, 10, 32)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse consolidations count: %v", err)
-		}
-		result[strings.TrimSpace(splits[1])] = int(countI64)
-	}
-	if err := scanner.Err(); err != nil && !errors.Is(err, io.EOF) {
-		return nil, fmt.Errorf("failed to read consolidations: %v", err)
-	}
-
-	return result, nil
 }
 
 // WaitForStatus waits till desired status of tablet is reached
@@ -400,10 +357,12 @@ func (vttablet *VttabletProcess) TearDownWithTimeout(timeout time.Duration) erro
 		return nil
 
 	case <-time.After(timeout):
-		vttablet.proc.Process.Kill()
-		err := <-vttablet.exit
-		vttablet.proc = nil
-		return err
+		proc := vttablet.proc
+		if proc != nil {
+			vttablet.proc.Process.Kill()
+			vttablet.proc = nil
+		}
+		return <-vttablet.exit
 	}
 }
 
@@ -455,28 +414,8 @@ func (vttablet *VttabletProcess) QueryTabletWithDB(query string, dbname string) 
 	return executeQuery(conn, query)
 }
 
-// executeQuery will retry the query up to 10 times with a small sleep in between each try.
-// This allows the tests to be more robust in the face of transient failures.
 func executeQuery(dbConn *mysql.Conn, query string) (*sqltypes.Result, error) {
-	var (
-		err    error
-		result *sqltypes.Result
-	)
-	retries := 10
-	retryDelay := 1 * time.Second
-	for i := 0; i < retries; i++ {
-		if i > 0 {
-			// We only audit from 2nd attempt and onwards, otherwise this is just too verbose.
-			log.Infof("Executing query %s (attempt %d of %d)", query, (i + 1), retries)
-		}
-		result, err = dbConn.ExecuteFetch(query, 10000, true)
-		if err == nil {
-			break
-		}
-		time.Sleep(retryDelay)
-	}
-
-	return result, err
+	return dbConn.ExecuteFetch(query, 10000, true)
 }
 
 // GetDBVar returns first matching database variable's value
@@ -526,7 +465,7 @@ func (vttablet *VttabletProcess) WaitForVReplicationToCatchup(t testing.TB, work
 		for duration > 0 {
 			log.Infof("Executing query %s on %s", query, vttablet.Name)
 			lastChecked = time.Now()
-			qr, err := executeQuery(conn, query)
+			qr, err := conn.ExecuteFetch(query, 1000, true)
 			if err != nil {
 				t.Fatal(err)
 			}
@@ -575,7 +514,7 @@ func (vttablet *VttabletProcess) BulkLoad(t testing.TB, db, table string, bulkIn
 	defer conn.Close()
 
 	query := fmt.Sprintf("LOAD DATA INFILE '%s' INTO TABLE `%s` FIELDS TERMINATED BY ',' ENCLOSED BY '\"'", tmpbulk.Name(), table)
-	_, err = executeQuery(conn, query)
+	_, err = conn.ExecuteFetch(query, 1, false)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -629,7 +568,6 @@ func VttabletProcessInstance(port, grpcPort, tabletUID int, cell, shard, keyspac
 	vttablet.VerifyURL = fmt.Sprintf("http://%s:%d/debug/vars", hostname, port)
 	vttablet.QueryzURL = fmt.Sprintf("http://%s:%d/queryz", hostname, port)
 	vttablet.StatusDetailsURL = fmt.Sprintf("http://%s:%d/debug/status_details", hostname, port)
-	vttablet.ConsolidationsURL = fmt.Sprintf("http://%s:%d/debug/consolidations", hostname, port)
 
 	return vttablet
 }
